bot/requirements.txt
discord.py>=2.4
httpx
python-dotenv
Pillow
